More file moving and organizing
